requirement.txt
python-dotenv
mysql-connector-python
google-generativeai
tk
